refactored folder structure to follow python convention
