+++ bin/generate_docs.py
