+++ docs/pbmc.html
@@ -391,6 +391,20 @@
               
               
               <tr>
+                  <td><a href="participant.html#cimac_participant_id">participant.cimac_participant_id</a></td>
+                  <td>Participant identifier assigned by the CIMAC-CIDC Network. Example: CIMAC-1234. </td>
+                  <td width=40%>
+                      &lt;string&gt; 
+                      <ul>
+                      
+                      </ul>
+                  </td>
+                </tr>
+            
+              
+              
+              
+              <tr>
                   <td><a href="participant.html#cohort_id">participant.cohort_id</a></td>
                   <td>Identifies the treatment cohort to which the participant belongs. Example: Cohort A. </td>
                   <td width=40%>
@@ -421,6 +435,20 @@
               <tr>
                   <td><a href="sample.html#site_sample_id">sample.site_sample_id</a></td>
                   <td>Sample identifier assigned by the biorepository site. Example: EA123-45. </td>
+                  <td width=40%>
+                      &lt;string&gt; 
+                      <ul>
+                      
+                      </ul>
+                  </td>
+                </tr>
+            
+              
+              
+              
+              <tr>
+                  <td><a href="sample.html#cimac_sample_id">sample.cimac_sample_id</a></td>
+                  <td>Specimen identifier assigned by the CIMAC-CIDC Network. Example: CIMAC-12453. </td>
                   <td width=40%>
                       &lt;string&gt; 
                       <ul>
